refactor: move app-related UI components to app/
BREAKING CHANGE: Module paths changed.
